kafka-consumer/requirements.txt
dispatch-py==0.7.0
kafka-python==2.0.2
requests==2.31.0
